Change GitHub API from v2 to v3
As described [on the GitHub API page][1], v2 is no longer supported, so the
code was modified to use v3 instead.

The implore function didn't work for me (I'm not sure whether this is because
the GitHub API switch), so I replaced it with a loop.

This commit fixes https://github.com/enygma/offline-issues/issues/2.

[1]: http://developer.github.com#june-12-2012